GS2_LiDAR_Distribution/Visualizer/requirements.txt
matplotlib>=3.5.0
pyserial>=3.5
numpy>=1.21.0
